Rename pkg/openai to pkg/runtime and fix template to use FixOpenAI
Rename the package to better reflect its broader scope beyond just OpenAI compatibility.
Update the generator template to use the comprehensive FixOpenAI function instead of
just FixMap, ensuring all OpenAI transformations are applied consistently.
