+++ pkg/generator/compatibility_test.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	. "github.com/onsi/gomega"
-	"github.com/redpanda-data/protoc-gen-go-mcp/pkg/openai"
+	"github.com/redpanda-data/protoc-gen-go-mcp/pkg/runtime"
 	"github.com/santhosh-tekuri/jsonschema/v5"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
@@ -333,7 +333,7 @@ func TestCompatOpenAI(t *testing.T) {
 			// Then apply Fix for OpenAI compatibility mode (converts all OpenAI format back to protobuf format)
 			// This is what would happen in the actual MCP tool execution before protojson.Unmarshal
 			if rawData, ok := jsonData.(map[string]interface{}); ok {
-				openai.Fix(tt.input.ProtoReflect().Descriptor(), rawData)
+				runtime.FixOpenAI(tt.input.ProtoReflect().Descriptor(), rawData)
 				
 				// Verify that the converted data can be unmarshaled into the proto message
 				fixedJSON, err := json.Marshal(rawData)
